+++ src/engine/phraseanditerator.cpp
@@ -30,7 +30,7 @@ PhraseAndIterator::PhraseAndIterator(const QVector<PostingIterator*>& iterators)
 {
 }
 
-uint PhraseAndIterator::docId()
+quint64 PhraseAndIterator::docId()
 {
     return m_docId;
 }
@@ -77,7 +77,7 @@ bool PhraseAndIterator::checkIfPositionsMatch()
     return !vec.isEmpty();
 }
 
-uint PhraseAndIterator::next()
+quint64 PhraseAndIterator::next()
 {
     if (m_iterators.isEmpty()) {
         m_docId = 0;
